refactor: refactor the UsersController tests
